+++ packages/core/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.1](https://github.com/medly/medly-components/compare/@medly-components/core@2.6.0...@medly-components/core@2.6.1) (2021-07-28)
+
+**Note:** Version bump only for package @medly-components/core
+
+
+
+
+
+# [2.6.0](https://github.com/medly/medly-components/compare/@medly-components/core@2.5.5...@medly-components/core@2.6.0) (2021-07-28)
+
+
+### Features
+
+* updated border radius ([1bda1cc](https://github.com/medly/medly-components/commit/1bda1ccf0fd25c6016e76249e566b7b68c4cd0de))
+
+
+
+
+
+## [2.5.5](https://github.com/medly/medly-components/compare/@medly-components/core@2.5.4...@medly-components/core@2.5.5) (2021-07-28)
+
+
+### Bug Fixes
+
+* **core:** fix grouped content cell right border issue ([dda9470](https://github.com/medly/medly-components/commit/dda9470e7ebc4491748dc3e1177a2cbd1dae9096))
+
+
+
+
+
 ## [2.5.4](https://github.com/medly/medly-components/compare/@medly-components/core@2.5.3...@medly-components/core@2.5.4) (2021-07-27)
 
 **Note:** Version bump only for package @medly-components/core